+++ src/FluentCertificates.Builder/CertificateSigningRequest.cs
@@ -154,7 +154,7 @@ public record CertificateSigningRequest
                 .ReadSequence(new Asn1Tag(TagClass.ContextSpecific, 0))
                 .ReadSequence(Asn1Tag.Sequence)
                 .ReadObjectIdentifier(Asn1Tag.ObjectIdentifier);
-            return SignatureAlgorithm.ForRsaSsaPss(algorithm, hashAlgorithm);
+            return SignatureAlgorithm.CreateForRsaSsaPss(algorithm, hashAlgorithm);
         }
         return SignatureAlgorithm.FromOidValue(algorithm);
     }